+++ R/createScaleInfo.R
@@ -2,15 +2,36 @@
 #############################################################################
 #' Create scale information.
 #'
-#' Create information on scale aggregation. Scales, 'fake scales', and imputed variables are listed.
+#' Create information on scale or imputation aggregation.
+#' Scales, 'fake scales', and imputed variables are listed.
+#'
+#' Currently, displaying the items of a imputed scale is not supported by the \code{codebook()} function.
+#' Even though the items of an imputed scale can be defined in the \code{scaleInfo},
+#' this information is simply ignored during the creation of the codebook.
 #'
 #'@param inputForDescriptives \code{inputForDescriptives} object as created by the \code{createInputForDescriptives} function
-#         source character string of length 1: source information of the data, for example \code{"sfb"}, \code{"lfb"}, etc.
 #'
-#'@return Returns the scale information.
+#'@return Returns a \code{data.frame} with the following information:
+#'\itemize{
+#'  \item \code{varName} The name of the variable as it occurs in the data
+#'  \item \code{Quelle} The name of the data set the variable belongs to.
+#'  This is only meaningful, if \code{inputForDescriptives} is a list
+#'  (meaning that the function runs for multiple data sets).
+#'  \item \code{Anzahl_valider_Werte} For a scale, how many values have to be not NA on items for a
+#'  non NA value on the scale?
+#'  \item \code{Items_der_Skala} Which items or imputations belong to the scale?
+#'  This column is used to display the item names in the codebook.
+#'  \item \code{Imputationen} Which imputations belong to the pooled variable?
+#'  This column is used to display the number of imputations in the codebook.
+#'}
 #'
 #'@examples
-#'#tbd
+#'# import spss exemplary data
+#'file <- system.file("extdata", "example1_clean.sav", package = "eatCodebook")
+#'dat   <- eatGADS::import_spss(file)
+#'inputForDescriptives <- createInputForDescriptives(dat, impExpr = "plausible value")
+#'
+#'scaleInfo <- createScaleInfo(inputForDescriptives)
 #'
 #'@export
 createScaleInfo <- function(inputForDescriptives){
@@ -18,21 +39,17 @@ createScaleInfo <- function(inputForDescriptives){
 }
 #'@export
 createScaleInfo.data.frame <- function(inputForDescriptives){
-#  if (!missing(source)) {
-#      if(!is.character(source) || length(source) != 1) {stop("'source' needs to be a character of length 1.")}
-#  } else {
-  source <- NA
-#  }
+  inputForDescriptives <- check_inputForDescriptives(inputForDescriptives)
 
+  source <- NA
   scaleInfo <- data.frame(varName = character(),	Quelle = character(),
-                          Anzahl_valider_Werte = character(),	Items_der_Skala = character(), stringsAsFactors = FALSE)
+                          Anzahl_valider_Werte = character(),
+                          Items_der_Skala = character(), Imputationen = character(),
+                          stringsAsFactors = FALSE)
 
   ## Scales
-  #browser()
-  inputForDescriptives <- check_inputForDescriptives(inputForDescriptives)
   scales_inputForDescriptives <- inputForDescriptives[which(inputForDescriptives$type %in% c("scale", "fake_item")), ]
   scales <- scales_inputForDescriptives[!duplicated(scales_inputForDescriptives$group), "group"]
-
   ## Inputed
   imputed_inputForDescriptives <- inputForDescriptives[which(inputForDescriptives$imp), ]
   pooled_variables <- unique(imputed_inputForDescriptives$group)
@@ -41,20 +58,18 @@ createScaleInfo.data.frame <- function(inputForDescriptives){
 
   for(i in seq_along(scales_and_imputed)) {
     single_inputForDescriptives <- inputForDescriptives[inputForDescriptives$group == scales_and_imputed[i], ]
-    # special treatment for imputed scales
-    if(all(single_inputForDescriptives$imp) && any(single_inputForDescriptives$type == "scale")) {
-      #browser()
-      scaleInfo[i, "varName"] <- unique(single_inputForDescriptives$group)
-      scaleInfo[i, "Items_der_Skala"] <- paste(single_inputForDescriptives$varName, collapse = ",")
-      scaleInfo[i, "Anzahl_valider_Werte"] <- "-"
-      scaleInfo[i, "Quelle"] <- source
-      next
-    }
     scaleInfo[i, "varName"] <- unique(single_inputForDescriptives$group)
-    items <- single_inputForDescriptives[single_inputForDescriptives$type %in% c("item", "fake_item"), "varName"]
-    scaleInfo[i, "Items_der_Skala"] <- paste(items, collapse = ",")
     scaleInfo[i, "Anzahl_valider_Werte"] <- "-"
     scaleInfo[i, "Quelle"] <- source
+
+    # scales
+    items <- single_inputForDescriptives[single_inputForDescriptives$type %in% c("item", "fake_item"), "varName"]
+    scaleInfo[i, "Items_der_Skala"] <- paste(items, collapse = ",")
+
+    #browser()
+    # imputations
+    imputations <- single_inputForDescriptives[single_inputForDescriptives$imp, "varName"]
+    scaleInfo[i, "Imputationen"] <- paste(imputations, collapse = ",")
   }
 
   scaleInfo