+++ NEWS.md
@@ -1,5 +1,6 @@
 # eatCodebook 0.2.0.9000
+* `createScaleInfo()` now treats scales and imputed variables separately via the columns 'Anzahl_der_Items' and 'Imputationen' (#28)
+
 
 # eatCodebook 0.2.0
-
 * Initial Github release.